feat: rewrite the definition of fuse.js
