--- definitions/npm/fuse.js_v6.x.x/test_fuse.js_v6.js
@@ -1,12 +1,0 @@
-// @flow
-import Fuse from "fuse.js";
-
-const fuse: Fuse<number> = new Fuse([1, 2]);
-fuse.search("foo");
-
-// $FlowExpectedError
-const A: number[] = fuse.search(1);
-
-fuse.setCollection([1, 2, 3, 4]);
-// $FlowExpectedError
-fuse.setCollection(1);
